requirements.txt
annotated-types==0.7.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
SQLAlchemy==2.0.30
typing_extensions==4.12.2
Werkzeug==3.0.3
